+++ typescript/infra/config/environments/testnet4/agent.ts
@@ -1,5 +1,4 @@
 import {
-  Chains,
   GasPaymentEnforcementPolicyType,
   RpcConsensusType,
   chainMetadata,
@@ -107,28 +106,7 @@ const releaseCandidate: RootAgentConfig = {
   },
 };
 
-const neutron: RootAgentConfig = {
-  ...contextBase,
-  context: Contexts.Neutron,
-  rolesWithKeys: [Role.Relayer],
-  contextChainNames: {
-    relayer: [Chains.goerli],
-    validator: [],
-    scraper: [],
-  },
-  relayer: {
-    rpcConsensusType: RpcConsensusType.Fallback,
-    docker: {
-      repo,
-      tag: '34611f0-20231218-204504',
-    },
-    gasPaymentEnforcement,
-    transactionGasLimit: 750000,
-  },
-};
-
 export const agents = {
   [Contexts.Hyperlane]: hyperlane,
   [Contexts.ReleaseCandidate]: releaseCandidate,
-  [Contexts.Neutron]: neutron,
 };